refactor(metal): move the FA-vec decode kernels out of attn.rs into fa_vec_decode.rs, no behaviour change
attn.rs 7999 -> 7337 lines. The four per-width kernel sources, the
threadgroup rule, the width predicate and the encoder move together.
